Move default-config into config/ module
